Decode integers to better fitting data types

+++ Examples/Pkl.ConsoleAppExample/Config/AppConfig.pkl.cs
@@ -4,7 +4,7 @@ public class AppConfig
 {
     public string Host { get; set; } = default!;
 
-    public int Port { get; set; } = default!;
+    public ushort Port { get; set; } = default!;
 
     public NestedConfig Nested { get; set; } = default!;
 }
